AMBARI-967. Enhance predicate comparison. (Tom Beerbower via mahadev)
git-svn-id: https://svn.apache.org/repos/asf/incubator/ambari/branches/AMBARI-666@1406489 13f79535-47bb-0310-9956-ffa450edef68

+++ ambari-server/src/main/java/org/apache/ambari/server/agent/ExecutionCommand.java
@@ -41,7 +41,7 @@ public class ExecutionCommand extends AgentCommand {
     super(AgentCommandType.EXECUTION_COMMAND);
   }
   private String clusterName;
-  private int taskId;
+  private long taskId;
   private String commandId;
   private String hostname;
   private Role role;
@@ -91,12 +91,12 @@ public class ExecutionCommand extends AgentCommand {
   }
 
   @JsonProperty("taskId")
-  public int getTaskId() {
+  public long getTaskId() {
     return taskId;
   }
 
   @JsonProperty("taskId")
-  public void setTaskId(int taskId) {
+  public void setTaskId(long taskId) {
     this.taskId = taskId;
   }
 
@@ -189,4 +189,5 @@ public class ExecutionCommand extends AgentCommand {
   public void setServiceName(String serviceName) {
     this.serviceName = serviceName;
   }
+
 }
